Move fetcher implementations into their own file
